Memory is now a struct + some initial work on brk
The "store_to_memory" and "load_from_memory" functions are now
implemented on a Memory struct, this was introduced so we can in the
future also store heap memory in this same struct and provide a
uniform interface to access both heap and stack memory in the same
way.

The syscall handling was changed a bit to use a match and some initial
work was done on supporting the brk syscall. It only changes the
"program_break" variable currently, it doesn't allocate any heap
memory at the moment.