Added/updated instructions & added address asserts
Added "sraiw" and fixed "srli" and "slriw" to use a logical shift
instead of using an arithmetic shift. Also added the "ori"
instruction.

Added asserts to memory access functions that use the new
is_address_valid function to check if the access will be valid or if
it will trigger an out of bounds exception.