Fixed label comment bug and updated instructions
Labels with comments after them now no longer break because the spacing
behind them was not trimmed away.

Added the "and", "bge" and "lb" instructions, aliased "mul" as "mulw",
updated slli/slliw to use only 5 bits to shift and use a logical shift
as described in the riscv spec. Aliased "lwu" as "lw" and "ld".