Initial work on supporting the .text directive
When .text is used somewhere after a label will result in the string
being stored at the very top of the stack. Any usage of that label will
just like with jump labels be replaced with the address of in this case
the string. To achieve this some changes had to be made in the compile
function to give the function access to the memory system.

The unescape library was added to unescape strings from the assembly
file.

The lla instruction was aliased into li, because the address pretty much
turns into an immediate value. Fixes were done to lbu and lb so that
they just use memory[address] instead of tyring to load 4 bytes and then
setting the remaining 3 to be 0 and/or then posibly sign extending it.
This change was made because it's more efficient and also because it
fixes issues where trying to load in byte 4095 when the virtual memory
size is 4096 would cause a crash because it would under the hood load
byte 4095 but also byte 4096, 4097 and 4098 all 3 of which are obviously
out of bounds.